core/server/requirements.txt
flask
openai
pyyaml
httpx
uvicorn
pytz
fastapi
paho-mqtt
zhdate
loguru
librosa
tqdm
pydub
resampy
opencv-python
soundfile
aiohttp
aiohttp_cors
aiortc
av
flask-sockets
fastapi
uvicorn
sqlalchemy
tiktoken
fuzzywuzzy
torch-ema
redis
ninja
trimesh
opencv-python
tensorboardX
numpy 
pandas
tqdm
matplotlib
PyMCubes
rich
dearpygui
packaging
scipy
scikit-learn

face_alignment
python_speech_features
numba
resampy
#pyaudio
soundfile==0.12.1
einops
configargparse

lpips==0.1.3
imageio-ffmpeg
numpy==1.24.4

transformers
edge_tts
flask
flask_sockets
opencv-python-headless
aiortc
aiohttp_cors

ffmpeg-python
omegaconf
diffusers
accelerate

librosa
openai
